+++ CHANGELOG.md
@@ -32,6 +32,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - [Refactor: Use Query-Parameters for Where-Clause](https://github.com/5pm-HDH/churchtools-api/pull/106)
 - [Fix Unit-Test for Person Update-Request](https://github.com/5pm-HDH/churchtools-api/pull/109)
 - [Fix Integration-Test: Cache Api-key](https://github.com/5pm-HDH/churchtools-api/pull/113)
+- [Abstract Model](https://github.com/5pm-HDH/churchtools-api/pull/118)
 
 ### Fixed